+++ evoacme/files/evoacme.sh
@@ -55,10 +55,8 @@ sed_cert_path_for_apache() {
 
         sed -i "s~${search}~${replace}~" "${vhost_full_path}"
         debug "Config in ${vhost_full_path} has been updated"
-        $(command -v apache2ctl) -t 2>&1 | grep -v "Syntax OK"
-        if [ "${PIPESTATUS[0]}" != "0" ]; then
-            error "Apache config test has exited with a non-zero exit code"
-        fi
+        $(command -v apache2ctl) -t 2>/dev/null
+        [ "${?}" -eq 0 ] || $(command -v apache2ctl) -t
     fi
 }
 sed_cert_path_for_nginx() {
@@ -76,7 +74,8 @@ sed_cert_path_for_nginx() {
 
         sed -i "s~${search}~${replace}~" "${vhost_full_path}"
         debug "Config in ${vhost_full_path} has been updated"
-        $(command -v nginx) -t
+        $(command -v nginx) -t 2>/dev/null
+        [ "${?}" -eq 0 ] || $(command -v nginx) -t
     fi
 }
 x509_verify() {